+++ app/src/main/java/host/stjin/anonaddy/ui/MainActivity.kt
@@ -230,6 +230,7 @@ class MainActivity : BaseActivity(), AddApiBottomDialogFragment.AddApiBottomDial
                     checkForSubscriptionExpiration()
                     checkForNewFailedDeliveries()
                     checkForNewAccountNotifications()
+                    ServiceLocator.aliasSearchManager.syncAllAliasesIfNeeded()
                 }
             }